# docs/swagger.yaml
basePath: /api
definitions:
...
  models.AuthResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.Company:
    properties:
      accounting_regime:
        type: string
      address:
        type: string
      annual_revenue:
        description: Dados operacionais
        type: number
      bank_name:
        description: Informação bancária
        type: string
      bic:
        type: string
      business_activity:
        type: string
      cae:
        type: string
      city:
        type: string
      company_name:
        description: Dados copiados da RegistrationRequest após aprovação
        type: string
      corporate_object:
        type: string
      country:
        type: string
      county:
        type: string
      created_at:
        type: string
      district:
        type: string
      estimated_revenue:
        type: number
      founding_date:
        type: string
      group_start_date:
        type: string
      has_stock:
        type: boolean
      iban:
        type: string
      id:
        type: integer
      legal_form:
        type: string
      main_clients:
        type: string
      main_suppliers:
        type: string
      monthly_invoices:
        type: integer
      nipc:
        type: string
      number_employees:
        type: integer
      postal_code:
        type: string
      registration_request:
        $ref: '#/definitions/models.RegistrationRequest'
      share_capital:
        type: number
      status:
        type: string
      trade_name:
        description: Dados completos
        type: string
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relacionamentos
      user_id:
        description: One-to-One com User
        type: integer
      vat_regime:
        type: string
    type: object
  models.CompleteCompanyDataDTO:
    properties:
      address:
        example: Rua das Flores, 123
        type: string
      annual_revenue:
        example: 100000
        type: number
      bank_name:
        example: Banco Comercial Português
        type: string
      bic:
        example: BCOMPTPL
        type: string
      city:
        example: Lisboa
        type: string
      corporate_object:
        example: Prestação de serviços de consultoria
        type: string
      county:
        example: Lisboa
        type: string
      district:
        example: Lisboa
        type: string
      group_start_date:
        example: "2024-01-01"
        type: string
      has_stock:
        example: false
        type: boolean
      iban:
        example: PT50000201231234567890154
        type: string
      main_clients:
        example: Cliente A, Cliente B
        type: string
      main_suppliers:
        example: Fornecedor X, Fornecedor Y
        type: string
      postal_code:
        example: 1000-001
        type: string
      share_capital:
        example: 5000
        type: number
      trade_name:
        example: Silva Consultoria
        type: string
    type: object
  models.CompleteUserDataDTO:
    properties:
      billing_software:
        example: Moloni
        type: string
      citizen_card_expiry:
        example: "2030-12-31"
        type: string
      citizen_card_number:
        example: "12345678"
        type: string
      e_fatura_password:
        example: password123
        type: string
      e_fatura_user:
        example: "123456789"
        type: string
      fiscal_county:
        example: Lisboa
        type: string
      fiscal_district:
        example: Lisboa
        type: string
      fixed_phone:
        example: "213456789"
        type: string
      marital_status:
        example: Solteiro
        type: string
      official_email:
        example: geral@empresa.com
        type: string
      portal_financas_password:
        example: password123
        type: string
      portal_financas_user:
        example: "123456789"
        type: string
      preferred_contact_hours:
        example: 9h-17h
        type: string
      preferred_format:
        example: digital
        type: string
      ss_direct_password:
        example: password123
        type: string
      ss_direct_user:
        example: "123456789"
        type: string
    type: object
  models.LoginRequest:
    properties:
      password:
        example: "123456"
        minLength: 6
        type: string
      username:
        example: joao.silva
        type: string
    required:
    - password
    - username
    type: object
  models.RegisterRequest:
    properties:
      email:
        example: joao@exemplo.com
        type: string
      name:
        example: João Silva
        minLength: 2
        type: string
      nif:
        example: "123456789"
        type: string
      password:
        example: "123456"
        minLength: 6
        type: string
      phone:
        example: "912345678"
        type: string
      role:
        enum:
        - client
        - accountant
        - admin
        example: client
        type: string
      status:
        enum:
        - pending
        - approved
        - rejected
        - blocked
        example: pending
        type: string
      username:
        example: joao.silva
        type: string
    required:
    - email
    - name
    - nif
    - password
    - phone
    - role
    - status
    - username
    type: object
  models.RegistrationRequest:
    properties:
      accounting_regime:
        type: string
      annual_revenue:
        description: Dados operacionais
        type: number
      approval_token:
        type: string
      bank_name:
        description: Informação bancária
        type: string
      bic:
        type: string
      billing_software:
        type: string
      business_activity:
        type: string
      cae:
        description: Dados básicos opcionais
        type: string
      citizen_card_expiry:
        type: string
      citizen_card_number:
        type: string
      company:
        $ref: '#/definitions/models.Company'
      company_address:
        type: string
      company_city:
        type: string
      company_country:
        type: string
      company_county:
        type: string
      company_district:
        type: string
      company_id:
        description: NULL até aprovação
        type: integer
      company_name:
        description: |-
          === DADOS DA COMPANY (armazenados até aprovação) ===
          Dados básicos obrigatórios
        type: string
      company_postal_code:
        type: string
      corporate_object:
        type: string
      created_at:
        type: string
      date_of_birth:
        description: Dados pessoais opcionais
        type: string
      email:
        type: string
      estimated_revenue:
        type: number
      fiscal_address:
        description: Morada fiscal
        type: string
      fiscal_city:
        type: string
      fiscal_county:
        type: string
      fiscal_district:
        type: string
      fiscal_postal_code:
        type: string
      fixed_phone:
        type: string
      founding_date:
        type: string
      group_start_date:
        type: string
      has_stock:
        type: boolean
      iban:
        type: string
      id:
        type: integer
      legal_form:
        type: string
      main_clients:
        type: string
      main_suppliers:
        type: string
      marital_status:
        type: string
      monthly_invoices:
        type: integer
      name:
        type: string
      nif:
        type: string
      nipc:
        type: string
      number_employees:
        type: integer
      official_email:
        description: Preferências
        type: string
      phone:
        type: string
      preferred_contact_hours:
        type: string
      preferred_format:
        type: string
      report_frequency:
        type: string
      request_type:
        description: new_client, existing_client
        type: string
      review_notes:
        type: string
      reviewed_at:
        type: string
      reviewed_by:
        type: integer
      reviewed_by_user:
        $ref: '#/definitions/models.User'
      share_capital:
        type: number
      status:
        type: string
      submitted_at:
        type: string
      tax_residence_country:
        type: string
      trade_name:
        description: Dados completos opcionais
        type: string
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relacionamentos
      user_id:
        description: === RELACIONAMENTOS (só existem após aprovação) ===
        type: integer
      username:
        description: |-
          === DADOS DO USER (armazenados até aprovação) ===
          Dados pessoais obrigatórios
        type: string
      vat_regime:
        type: string
    type: object
  models.RegistrationRequestDTO:
    properties:
      accounting_regime:
        example: organizada
        type: string
      annual_revenue:
        example: 100000
        type: number
      bank_name:
        example: Banco Comercial Português
        type: string
      bic:
        example: BCOMPTPL
        type: string
      billing_software:
        example: Moloni
        type: string
      business_activity:
        example: Consultoria em gestão
        type: string
      cae:
        example: "69200"
        type: string
      citizen_card_expiry:
        example: "2030-12-31"
        type: string
      citizen_card_number:
        example: "12345678"
        type: string
      company_address:
        example: Rua das Flores, 123
        type: string
      company_city:
        example: Lisboa
        type: string
      company_country:
        example: Portugal
        type: string
      company_county:
        example: Lisboa
        type: string
      company_district:
        example: Lisboa
        type: string
      company_name:
        description: |-
          === DADOS EMPRESA ===
          Obrigatórios
        example: Silva & Associados Lda
        type: string
      company_postal_code:
        example: 1000-001
        type: string
      corporate_object:
        example: Prestação de serviços de consultoria
        type: string
      date_of_birth:
        description: === TODOS OS CAMPOS OPCIONAIS ===
        example: "1990-01-15"
        type: string
      email:
        example: joao@exemplo.com
        type: string
      estimated_revenue:
        example: 50000
        type: number
      fiscal_address:
        description: Morada fiscal obrigatória
        example: Rua das Flores, 123
        type: string
      fiscal_city:
        example: Lisboa
        type: string
      fiscal_county:
        example: Lisboa
        type: string
      fiscal_district:
        example: Lisboa
        type: string
      fiscal_postal_code:
        example: 1000-001
        type: string
      fixed_phone:
        example: "213456789"
        type: string
      founding_date:
        example: "2024-01-15"
        type: string
      group_start_date:
        example: "2024-01-01"
        type: string
      has_stock:
        example: false
        type: boolean
      iban:
        example: PT50000201231234567890154
        type: string
      legal_form:
        example: Sociedade por Quotas
        type: string
      main_clients:
        example: Cliente A, Cliente B
        type: string
      main_suppliers:
        example: Fornecedor X, Fornecedor Y
        type: string
      marital_status:
        example: Solteiro
        type: string
      monthly_invoices:
        example: 10
        type: integer
      name:
        example: João Silva
        type: string
      nif:
        example: "123456789"
        type: string
      nipc:
        description: Opcionais
        example: "123456789"
        type: string
      number_employees:
        example: 2
        type: integer
      official_email:
        example: geral@empresa.com
        type: string
      password:
        example: password123
        minLength: 6
        type: string
      phone:
        example: "912345678"
        type: string
      preferred_contact_hours:
        example: 9h-17h
        type: string
      preferred_format:
        example: digital
        type: string
      report_frequency:
        example: mensal
        type: string
      share_capital:
        example: 5000
        type: number
      tax_residence_country:
        example: Portugal
        type: string
      trade_name:
        example: Silva Consultoria
        type: string
      username:
        description: |-
          === DADOS PESSOAIS ===
          Obrigatórios
        example: joao.silva
        type: string
      vat_regime:
        example: normal
        type: string
    required:
    - company_name
    - email
    - fiscal_address
    - fiscal_city
    - fiscal_postal_code
    - legal_form
    - name
    - nif
    - password
    - phone
    - username
    type: object
  models.SuccessResponse:
    properties:
      data: {}
      message:
        example: Operação realizada com sucesso
        type: string
      success:
        example: true
        type: boolean
    type: object
  models.UpdateCompanyDTO:
    properties:
      address:
        example: Rua das Flores, 123
        type: string
      city:
        example: Lisboa
        type: string
      postal_code:
        example: 1000-001
        type: string
      trade_name:
        example: Silva Consultoria
        type: string
    type: object
  models.UpdateProfileDTO:
    properties:
      name:
        example: João Silva
        type: string
      phone:
        example: "912345678"
        type: string
    type: object
...
  models.User:
    properties:
      billing_software:
        type: string
      citizen_card_expiry:
        type: string
      citizen_card_number:
        type: string
      company:
        allOf:
        - $ref: '#/definitions/models.Company'
        description: Relacionamentos
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      date_of_birth:
        description: Dados copiados da RegistrationRequest após aprovação
        type: string
      email:
        example: joao@exemplo.com
        type: string
      fiscal_address:
        description: Morada fiscal completa
        type: string
      fiscal_city:
        type: string
      fiscal_county:
        type: string
      fiscal_district:
        type: string
      fiscal_postal_code:
        type: string
      fixed_phone:
        description: Contactos adicionais
        type: string
      id:
        example: 1
        type: integer
      marital_status:
        type: string
      name:
        example: João Silva
        type: string
      nif:
        example: "123456789"
        type: string
      official_email:
        description: Acessos e credenciais (guardados encriptados)
        type: string
      phone:
        example: "912345678"
        type: string
      preferred_contact_hours:
        type: string
      preferred_format:
        description: Preferências de comunicação
        type: string
      registration_request:
        $ref: '#/definitions/models.RegistrationRequest'
      report_frequency:
        type: string
      reviewed_requests:
        items:
          $ref: '#/definitions/models.RegistrationRequest'
        type: array
      role:
        example: client
        type: string
      status:
        description: Sempre approved quando criado
        example: approved
        type: string
      tax_residence_country:
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      username:
        example: joao.silva
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Sistema de gestão contabilística com aprovação de clientes
  title: RV Contabilidade API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login com username e password
      parameters:
      - description: Username e password
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthResponse'
      summary: Entrar
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Remove token do lado do cliente
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Logout do utilizador
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Cria uma nova solicitação de registo para aprovação da contabilista
      parameters:
      - description: Dados de registo completos
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegistrationRequestDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SuccessResponse'
      summary: Registo de novo cliente
      tags:
      - auth
  /auth/register-direct:
    post:
      consumes:
      - application/json
      description: Regista novo utilizador diretamente (para uso interno)
      parameters:
      - description: Dados de registo
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AuthResponse'
      summary: Criar conta diretamente
      tags:
      - auth
  /client/company:
    get:
      consumes:
      - application/json
      description: Obtém dados da empresa do cliente logado
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Dados da empresa do cliente
      tags:
      - client
    put:
      consumes:
      - application/json
      description: Atualiza dados da empresa do cliente (campos limitados)
      parameters:
      - description: Dados a atualizar
        in: body
        name: company
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCompanyDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Atualizar dados da empresa
      tags:
      - client
  /client/complete-company-data:
    put:
      consumes:
      - application/json
      description: Completa dados da empresa adicionais após aprovação
      parameters:
      - description: Dados da empresa completos
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.CompleteCompanyDataDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Completar dados da empresa
      tags:
      - client
  /client/complete-user-data:
    put:
      consumes:
      - application/json
      description: Completa dados pessoais adicionais após aprovação
      parameters:
      - description: Dados pessoais completos
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.CompleteUserDataDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Completar dados pessoais
      tags:
      - client
  /client/profile:
    get:
      consumes:
      - application/json
      description: Obtém perfil do cliente logado
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Perfil do cliente
      tags:
      - client
    put:
      consumes:
      - application/json
      description: Atualiza dados pessoais do cliente
      parameters:
      - description: Dados a atualizar
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/models.UpdateProfileDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Atualizar perfil do cliente
      tags:
      - client
  /client/requests:
    get:
      consumes:
      - application/json
      description: Lista todas as solicitações do cliente logado
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Histórico de solicitações do cliente
      tags:
      - client
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"